requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.7
importlib_resources==6.4.4
itsdangerous==2.1.2
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.7
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.0
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-telegram-bot==21.5
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
ruff==0.6.2
safetensors==0.4.5
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
SQLAlchemy==2.0.29
starlette==0.38.2
streamlit==1.34.0
sympy==1.13.2
tenacity==8.3.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.2.1
uvicorn==0.30.6
watchdog==4.0.0
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
